fix: pathing after refactor
