+++ apps/router/src/guardian-ui/components/dashboard/tabs/meta/manager/MetaManager.tsx
@@ -11,10 +11,10 @@ import {
 } from '@chakra-ui/react';
 import { fieldsToMeta, metaToHex, useTranslation } from '@fedimint/utils';
 import { ParsedConsensusMeta } from '@fedimint/types';
-import { ModuleRpc } from '../../../../../types';
 import { DEFAULT_META_KEY } from '../../FederationTabsCard';
-import { useGuardianAdminApi } from '../../../../../../context/hooks';
 import { RequiredMeta } from './RequiredMeta';
+import { useGuardianAdminApi } from '../../../../../../context/hooks';
+import { ModuleRpc } from '../../../../../../types/guardian';
 
 const metaArrayToObject = (
   metaArray: [string, string][]
