+++ apps/router/src/guardian-ui/components/setup/SetupProgress.tsx
@@ -11,7 +11,7 @@ import {
 } from '@chakra-ui/react';
 import { ReactComponent as CheckIcon } from '../../assets/svgs/white-check.svg';
 import { useTranslation } from '@fedimint/utils';
-import { StepState } from '../../types';
+import { StepState } from '../../../types/guardian';
 
 interface StepProps {
   text: string;
