+++ apps/router/src/guardian-ui/components/setup/screens/connectGuardians/ConnectGuardians.tsx
@@ -17,13 +17,13 @@ import {
 import { CopyInput, Table, TableRow } from '@fedimint/ui';
 import { ModuleKind, GuardianServerStatus } from '@fedimint/types';
 import { useTranslation } from '@fedimint/utils';
-import { GuardianRole } from '../../../../types';
 import { getModuleParamsFromConfig } from '../../../../utils/api';
 import { ReactComponent as CopyIcon } from '../../../../assets/svgs/copy.svg';
 import {
   useConsensusPolling,
   useGuardianSetupContext,
 } from '../../../../../context/hooks';
+import { GuardianRole } from '../../../../../types/guardian';
 
 interface Props {
   next(): void;
